% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_content_type}
\alias{check_content_type}
\title{Checa o tipo de conteúdo da resposta da requisição}
\usage{
check_content_type(response)
}
\arguments{
\item{response}{Um objeto com o retorno (resposta) de uma chamada `httr:GET`}
}
\description{
Checa o tipo de conteúdo da resposta da requisição
}
