% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_id}
\alias{check_input_id}
\title{Checa o tipo do input de id}
\usage{
check_input_id(x)
}
\arguments{
\item{x}{um ID específico de cada uma das funções `get_`}
}
\description{
Checa o tipo do input de id
}
