% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_requisition}
\alias{check_requisition}
\title{Checa o codigo da requisicao}
\usage{
check_requisition(response)
}
\arguments{
\item{response}{Um objeto com o retorno (resposta) de uma chamada `httr:GET`}
}
\description{
Checa o codigo da requisicao
}
