% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font.R
\name{get_font}
\alias{get_font}
\title{Obtém informações sobre uma fonte de dados das séries da API.}
\usage{
get_font(font_id)
}
\arguments{
\item{font_id}{Um número inteiro que representa um ID válido de umas das fontes de dados da API.}
}
\value{
`get_font` retorna um data.frame com duas colunas:

 id: o número identificador da fonte de dados.

 titulo: uma string com o nome da fonte de dados.
}
\description{
`get_font` obtém informações de identificação de uma fonte dos dados das séries disponíveis na API.
}
\examples{

\dontrun{

data <- get_fonts(font_id = 5)

}

}
