% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periodicities.R
\name{get_periodicities}
\alias{get_periodicities}
\title{Obtém informações sobre a periodicidade dos dados.}
\usage{
get_periodicities()
}
\value{
`get_periodicities` retorna um data.frame com duas colunas:

 id: o número identificador da periodicidade das séries da API.

 titulo: uma string com o nome da periodicidade das séries da API.
}
\description{
`get_periodicities` Obtém informações sobre a periodicidade dos dados disponíveis na API.
}
\examples{

\dontrun{

data <- get_periodicities()

}

}
