% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_serie.R
\name{get_serie}
\alias{get_serie}
\title{Obtém informações de uma série através do ID da série.}
\usage{
get_serie(serie_id)
}
\arguments{
\item{serie_id}{Um número inteiro, indicando um ID único e válido de uma série disponibilizado na API.}
}
\value{
`get_serie` retorna um data.frame com duas variáveis:

 id: o número identificador da série (series_id).

 titulo: uma string com o título da série.
}
\description{
`get_serie` obtém informações de identificação de uma série específica através seu proprio ID. Para ter uma ma lista completa de todas as séries disponíveis e acessar seus IDs, você pode usar a função `get_themes` e encontrar todas as informações de séries por tema. Além disso, pode usar afunção `get_series` para buscar informações sobre uma série pelo ID de seu tema.
}
\examples{
\dontrun{

data <- get_serie(serie_id=1)

}

}
