% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Obtém informações de uma série através do ID de tema}
\usage{
get_series(theme_id)
}
\arguments{
\item{theme_id}{Um número inteiro, indicando um ID único e válido de um tema disponibilizado na API.}
}
\value{
`get_series` retorna um data.frame com duas variáveis:

 id: o número identificador da série (series_id).

 titulo: uma string com o título da série.
}
\description{
`get_series` obtém informações de identificação de uma série específica através do ID seu tema. Uma lista completa dos IDs de temas pode ser encontrada atráves da função `get_themes`.
}
\examples{
\dontrun{

data <- get_series(theme_id=1)

}

}
