% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series_values.R
\name{get_series_values}
\alias{get_series_values}
\title{Obtém os dados de uma série buscando pelo seu ID}
\usage{
get_series_values(serie_id, geographic_scope)
}
\arguments{
\item{serie_id}{Um número inteiro, indicando um ID único e válido de uma série disponibilizada na API.}

\item{geographic_scope}{Um número inteiro que indica a abrangencia geográfica da série: `1` para País; `2` para Região; `3` para UF (Estado) e `4` para Município.}
}
\value{
`get_series_values` retorna um data.frame com quatro colunas:

 cod: um númeo inteiro com o código do IBGE da Unidade Geográfica da informação.

 sigla: uma string com a abreviação do nome da Unidade Geográfica de acordo com o IBGE.

 valor: uma coluna com o valor da série na Unidade Geográfica. Frequentemente será um valor númerico.

 periodo: uma string com a informação acerca da data de referência para a série. Há três possibilidades de formato: YYYY-MM-DD, YYYY, YYYY-MM.
}
\description{
`get_series_values` obtém dados de uma série especificada pelo seu ID. Uma lista com IDs únicos de cada série pode ser buscada atráves da função `get_series`.
}
\examples{
\dontrun{

data <- get_series_values(serie_id=165, geographic_scope=3)

}

}
