% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theme.R
\name{get_theme}
\alias{get_theme}
\title{Obtém informações de um tema através do ID do tema.}
\usage{
get_theme(theme_id)
}
\arguments{
\item{theme_id}{Um número inteiro, indicando um ID único e válido de um tema disponibilizado na API.}
}
\value{
`get_theme` retorna um data.frame com quatro colunas:

 id: o número identificador do tema (theme_id).

 titulo: uma string com o título do tema.

 imagem: o nome da imagem do tema para uso das páginas web do Atlas da Violência.

 tipo: um identificador do tipo. NOTA: esta coluna está completamente preenchida com zeros (0). Pode ser alguma nova característica da API para futuros desenvolvimentos.
}
\description{
`get_theme` obtém informações de identificação de um tema disponível na API atráves de seu próprio ID. Você pode usar a função `get_themes` para buscar todos os temas e IDs disponíveis.
}
\examples{
\dontrun{

data <- get_themes()

}

}
