% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_themes.R
\name{get_themes}
\alias{get_themes}
\title{Obtém informações de uma série através do ID da série.}
\usage{
get_themes()
}
\value{
`get_themes` retorna um data.frame com seis colunas:

 id: o número identificador do tema (theme_id).

 titulo: uma string com o título do tema.

 tema_id: um identificador do tema. NOTA: esta coluna está completamente preenchida com zeros (0). Pode ser alguma nova característica da API para futuros desenvolvimentos.

 imagem: o nome da imagem do tema para uso das páginas web do Atlas da Violência.

 tipo: um identificador do tipo. NOTA: esta coluna está completamente preenchida com zeros (0). Pode ser alguma nova característica da API para futuros desenvolvimentos.

 subTemas: indica os subtemas, quando houver. NOTA: esta coluna está completamente preenchida com zeros (0). Pode ser alguma nova característica da API para futuros desenvolvimentos.
}
\description{
`get_themes` obtém informações de identificação das séries disponíveis na API por tipo de tema.
}
\examples{

\dontrun{

data <- get_themes()

}

}
