% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R, R/get_unit.R
\name{get_unit}
\alias{get_unit}
\title{Obtém informações sobre uma unidade de medida das séries da API por ID.}
\usage{
get_unit(unit_id)

get_unit(unit_id)
}
\arguments{
\item{unit_id}{Número inteiro com o ID de uma unidade de medida específica.}
}
\value{
`get_unit` retorna um data.frame com duas colunas:

 id: o número identificador da unidade de medida das séries da API.

 titulo: uma string com o nome da unidade de medida das séries da API.

`get_unit` retorna um data.frame com duas colunas:

 id: o número identificador da unidade de medida das séries da API.

 titulo: uma string com o nome da unidade de medida das séries da API.
}
\description{
`get_unit` obtém informações de identificação de uma unidade de medida das séries da API por ID.

`get_unit` obtém informações de identificação de uma unidade de medida das séries da API por ID.
}
\examples{

\dontrun{

data <- get_unit(unit_id=9)

}


\dontrun{

data <- get_unit(unit_id=9)

}

}
