% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units.R
\name{get_units}
\alias{get_units}
\title{Obtém informações sobre as unidades de medida de dados das séries da API.}
\usage{
get_units()
}
\value{
`get_units` retorna um data.frame com duas colunas:

 id: o número identificador da unidade de medida dos dados.

 titulo: uma string com o nome da da unidade de medida dos dados.
}
\description{
`get_units` obtém informações de identificação das unidades de medida dos dados das séries disponíveis na API.
}
\examples{

\dontrun{

data <- get_fonts()

}

}
